+++ src/Aerofly/MissionCheckpoint.ts
@@ -44,6 +44,10 @@ export class MissionCheckpoint {
    */
   frequency: number = 0;
   /**
+   * If waypoint is meant to be flown over. Else turn anticipation will be used.
+   */
+  flyOver: boolean = false;
+  /**
    * Not official: In kts TAS
    */
   speed: number = -1;
@@ -234,6 +238,11 @@ export class MissionCheckpoint {
   }
 
   toString(index: number): string {
+    let flyOver = "";
+    if (this.type === MissionCheckpoint.TYPE_WAYPOINT) {
+      flyOver = `                        <[bool][FlyOver][${this.flyOver ? "true" : "false"}]>
+      `;
+    }
     return `                    <[tmmission_checkpoint][element][${index}]
                         <[string8u][type][${Quote.tmc(this.type)}]>
                         <[string8u][name][${Quote.tmc(this.name)}]>
@@ -244,7 +253,15 @@ export class MissionCheckpoint {
                         <[float64][slope][${this.slope}]> // ${this.slope_deg.toFixed(1)} deg
                         <[float64][length][${this.length}]>
                         <[float64][frequency][${this.frequency.toFixed()}]>
-                    >
+${flyOver}                    >
+`;
+  }
+
+  toStringTargetPlane(name: string = "finish"): string {
+    return `                <[tmmission_target_plane][${name}][]
+                    <[vector2_float64][lon_lat][${this.lon_lat.lon} ${this.lon_lat.lat}]>
+                    <[float64][direction][${this.direction}]>
+                >
 `;
   }